hearts/tests: split off test_card
