+++ hearts/tests/test_card.py
@@ -1,0 +1,18 @@
+import pytest
+
+from hearts.hearts import Card
+
+
+def test_card_serialize():
+    card = Card('A', 'h')
+    assert 'Ah' == card.serialize()
+    assert card == Card.deserialize('Ah')
+
+
+def test_card_validate():
+    with pytest.raises(ValueError):
+        Card.deserialize('Ch')
+    with pytest.raises(ValueError):
+        Card.deserialize('Q4')
+    with pytest.raises(ValueError):
+        Card.deserialize('2t')
